+++ MadalinaBalaita/Imobile/Imobile.Core/Estates/House.cs
@@ -4,7 +4,7 @@ namespace Imobile.Core.Estates
    public class House :Estate
     {
         
-        public House( int sm1,string location1,string condition1) : base( sm1, location1,condition1)
+        public House(double sm1,string location1,string condition1) : base( sm1, location1,condition1)
         {
 
         }
