+++ MadalinaBalaita/Imobile/Imobile.Core/Estates/Studio.cs
@@ -3,7 +3,7 @@ namespace Imobile.Core.Estates
 {
     public class Studio : Estate
     {
-        public Studio( int sm1,  string location1,string condition1) : base(sm1, location1,condition1)
+        public Studio(double sm1,  string location1,string condition1) : base(sm1, location1,condition1)
         {
 
         }
